Drop support for legacy Windows requiring emulation
