+++ Contributing.md
@@ -12,8 +12,6 @@ The `ansi-terminal` and `ansi-terminal-types` packages aim to have:
 As of `ansi-terminal-0.11.5`, the package aims to be compatible with versions
 of GHC from GHC 7.10.1 (released March 2015). GHC 7.10.1 comes with:
 * `base-4.8.0.0`
-* `containers-0.5.6.2` (used for its `Data.Map.Strict.Map` and only in the
-  Windows version of the package)
 * `Win32-2.3.1.0` (only on Windows)
 
 The `ansi-terminal` package also depends on:
@@ -30,27 +28,19 @@ directives. The `CPP` GHC extension is specified for the libraries in the Cabal
 files. The `Win32` package has developed over time and module
 `System.Win32.Compat` is used to reduce the use of CPP pragmas in other modules.
 
-CPP `#include` pragmas are also used to include code from files to limit code
-duplication in modules that are specific to the 'Unix' or Windows versions of
-the package. The common code is located in folder `src\includes`.
-
-Separate 'Unix' and Windows versions of the package exist because, before
-Windows 10 version 1511, the native terminal software on Windows did not support
-the control sequences and emulation was required. Microsoft has not supported
-Windows XP since 8 April 2014 (and GHC has not supported Windows XP from
-GHC 8.0.1 of May 2016); Windows Server 2003 since 14 July 2015; Windows Vista
-since 11 April 2017; Windows 7, service pack 1 since 10 January 2023;
+Separate 'Unix' and Windows versions of the package used to exist because,
+before Windows 10 version 1511, the native terminal software on Windows did not
+support the control sequences and emulation was required. Microsoft has not
+supported Windows XP since 8 April 2014 (and GHC has not supported Windows XP
+from GHC 8.0.1 of May 2016); Windows Server 2003 since 14 July 2015; Windows
+Vista since 11 April 2017; Windows 7, service pack 1 since 10 January 2023;
 Windows 8.1 since 10 January 2023; and Windows 10, version 1507 (which 1511 of
-December 2015 updated) since 9 May 2017. Consequently, currently, there is no
-version of Windows that both has support by Microsoft and needs emulation. This
-may make changes to the emulation difficult to test.
+December 2015 updated) since 9 May 2017. Consequently, there is no longer any
+version of Windows that both has support by Microsoft and needs emulation.
 
-The package uses GHC's 'Safe Haskell' language extensions `Trustworthy`
-(introduced in GHC 7.2.1) and `Safe` (introduced in GHC 7.2.1 but not stable
-until GHC 7.4.1) to add flags explicitly to all modules. For most modules, that
-is done using a CPP pragma: `#include "Common-Safe-Haskell.hs"`. Modules
-`System.Console.ANSI.Windows.Detect` and `System.Console.ANSI.Windows.Emulator`
-use `System.IO.Unsafe.unsafePerformIO` but are flagged `Trustworthy`.
+The package uses GHC's 'Safe Haskell' language extension `Safe` (introduced in
+GHC 7.2.1 but not stable until GHC 7.4.1) to add flags explicitly to all
+modules.
 
 The source code generally follows
 [Johan Tibell's style guide](https://github.com/tibbe/haskell-style-guide), but
