--- ansi-terminal/src/includes/Common-Safe-Haskell.hs
@@ -1,1 +1,0 @@
-{-# LANGUAGE Safe #-}
